pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sculpt"
version = "0.1.35"
description = "Sculpt: Structuring unstructured data with LLMs"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Ben Turtel", email = "ben@lightningrod.ai" }
]
requires-python = ">=3.7"
keywords = [
    "data sculpting",
    "llm",
    "large language model",
    "unstructured data",
    "structured data",
    "data extraction",
    "information extraction",
    "data transformation",
    "text processing",
    "natural language processing",
    "nlp",
    "pipeline",
    "text to structured data",
]

classifiers = [
    "Development Status :: 3 - Alpha",  # Or 4 - Beta
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",  # Or your chosen license
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
]

dependencies = [
  "openai>=0.27.0",
  "tqdm>=4.0",
  "PyYAML>=6.0",
  "asgiref>=3.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://lightningrod.ai"
Source = "https://github.com/lightning-rod-labs/sculptor"
Repository = "https://github.com/lightning-rod-labs/sculptor.git"